Move test to module level for component test
