+++ src/features/suggestions/convert/index.test.ts
@@ -1,4 +1,4 @@
-import { parseUnit } from ".";
+import convert, { isMeasurementConvertible, parseUnit } from ".";
 
 test("convert standard unit", () => {
   const result = parseUnit("cup");
@@ -8,4 +8,32 @@ test("convert standard unit", () => {
 test("convert variable unit", () => {
   const result = parseUnit("of 12-inch");
   expect(result).toMatchObject({ diameter: 12 });
+})
+
+test("similar variable units e.g. diameter units are convertible", () => {
+  const result = isMeasurementConvertible({ diameter: 12 }, { unit: { diameter: 14 } });
+  expect(result).toBeTruthy();
+})
+
+test("similar standard units e.g. volume units are convertible", () => {
+  const result = isMeasurementConvertible("fl-oz", { unit: "cup" });
+  expect(result).toBeTruthy();
+})
+
+test("standard unit and variable unit are not convertible, vice versa", () => {
+  expect(isMeasurementConvertible("fl-oz", { unit: { diameter: 12 } })).toBeFalsy();
+  expect(isMeasurementConvertible({ diameter: 12 }, { unit: "fl-oz" })).toBeFalsy();
+})
+
+test("similar variable units e.g. diameter units should convert", () => {
+  expect(convert(1, { diameter: 2 }, { diameter: 1 })).toBe(4)
+})
+
+test("similar standard units e.g. volume units should convert", () => {
+  expect(convert(8, "fl-oz", "cup")).toBe(1)
+})
+
+test("standard unit and variable unit should return NaN, vice versa", () => {
+  expect(convert(1, "fl-oz", { diameter: 12 })).toBe(NaN);
+  expect(convert(1, { diameter: 12 }, "fl-oz")).toBe(NaN);
 })
